requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
dj-stripe==2.8.3
Django==4.2.6
django-cors-headers==4.3.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
environ==1.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
packaging==23.2
Pillow==10.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
sqlparse==0.4.4
stripe==4.2.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
